+++ src/ExploreHomer.vue
@@ -16,6 +16,8 @@
 
   import ReaderWidget from '@scaife-viewer/widget-reader';
   import MetadataWidget from '@scaife-viewer/widget-metadata';
+  // eslint-disable-next-line max-len
+  import PassageCitationTOCWidget from '@scaife-viewer/widget-passage-citation-toc';
   import PassageReferenceWidget from '@scaife-viewer/widget-passage-reference';
   import PassageAncestorsWidget from '@scaife-viewer/widget-passage-ancestors';
   import PassageSiblingsWidget from '@scaife-viewer/widget-passage-siblings';
@@ -52,6 +54,7 @@
         return [
           LibraryWidget,
           PassageReferenceWidget,
+          PassageCitationTOCWidget,
           PassageAncestorsWidget,
           PassageSiblingsWidget,
           PassageChildrenWidget,