+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Features
 
 - Add an [Attributions Widget](/packages/widget-attributions)
+- Add [Passage Citation Table of Contents Widget](packages/widget-passage-citation-toc)
 
 ## 0.2.0 (2020-11-11)